add log to observe instead of oom for issue 2638
