+++ src/main/java/com/actiontech/dble/route/impl/AbstractRouteStrategy.java
@@ -12,14 +12,42 @@ import com.actiontech.dble.route.util.RouterUtil;
 import com.actiontech.dble.server.parser.ServerParse;
 import com.actiontech.dble.services.mysqlsharding.ShardingService;
 import com.actiontech.dble.sqlengine.mpp.LoadData;
+import com.alibaba.druid.sql.ast.SQLStatement;
+import com.alibaba.druid.sql.dialect.mysql.parser.MySqlStatementParser;
+import com.alibaba.druid.sql.parser.SQLStatementParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.sql.SQLException;
+import java.sql.SQLSyntaxErrorException;
+import java.util.List;
 
 public abstract class AbstractRouteStrategy implements RouteStrategy {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractRouteStrategy.class);
+    protected static final Logger LOGGER = LoggerFactory.getLogger(DefaultRouteStrategy.class);
+
+    @Override
+    public SQLStatement parserSQL(String originSql) throws SQLSyntaxErrorException {
+        SQLStatementParser parser = new MySqlStatementParser(originSql);
+
+        /**
+         * thrown SQL SyntaxError if parser error
+         */
+        try {
+            List<SQLStatement> list = parser.parseStatementList();
+            if (list.size() > 1) {
+                throw new SQLSyntaxErrorException("MultiQueries is not supported,use single query instead ");
+            }
+            return list.get(0);
+        } catch (Exception t) {
+            LOGGER.warn("routeNormalSqlWithAST", t);
+            if (t.getMessage() != null) {
+                throw new SQLSyntaxErrorException(t.getMessage());
+            } else {
+                throw new SQLSyntaxErrorException(t);
+            }
+        }
+    }
 
     @Override
     public RouteResultset route(SchemaConfig schema, int sqlType, String origSQL,
@@ -44,7 +72,6 @@ public abstract class AbstractRouteStrategy implements RouteStrategy {
             if (sqlType == ServerParse.SHOW) {
                 rrs.setStatement(origSQL);
                 rrs = RouterUtil.routeToSingleNode(rrs, schema.getRandomShardingNode());
-                service.getSession2().endParse();
             } else {
                 rrs = routeNormalSqlWithAST(schema, origSQL, rrs, service, isExplain);
             }
@@ -54,12 +81,15 @@ public abstract class AbstractRouteStrategy implements RouteStrategy {
         return rrs;
     }
 
+    @Override
+    public RouteResultset route(SchemaConfig schema, int sqlType, String origSQL, ShardingService service) throws SQLException {
+        return this.route(schema, sqlType, origSQL, service, false);
+    }
 
     /**
      * routeNormalSqlWithAST
      */
-    public abstract RouteResultset routeNormalSqlWithAST(SchemaConfig schema, String stmt, RouteResultset rrs,
-                                                         ShardingService service, boolean isExplain) throws SQLException;
-
+    protected abstract RouteResultset routeNormalSqlWithAST(SchemaConfig schema, String stmt, RouteResultset rrs,
+                                                            ShardingService service, boolean isExplain) throws SQLException;
 
 }
